Centralização da url

+++ src/components/Listas/MembroId/index.jsx
@@ -1,33 +1,80 @@
+const url = import.meta.env.VITE_API_URL
+
 import { Link } from 'react-router-dom';
 import { formatarDataCriacaoAtualizacao } from '../../../utils/fomatters'
 import styles from './membro.module.css'
 import { HiOutlinePencilAlt } from "react-icons/hi";
-import { BsTrash3Fill } from "react-icons/bs";
-import { useEffect, useState } from 'react';
+import { BsTrash3Fill, BsTrashFill } from "react-icons/bs";
+import Swal from 'sweetalert2'
 
-export const MembroId = ({ membro, composicaoFamiliar }) => {
+export const MembroId = ({ membro, composicaoFamiliar, onRemoveComposicao, onRemoveMembro }) => {
 
-    const handleRemove = async (e) => {
-        e.preventDefault()
-        const response = await fetch(`http://localhost:3000/composicao/${composicaoFamiliar.id}`, {
-            method: 'DELETE',
-            credentials: 'include',
-            headers: {
-                'Content-Type': 'application/json'
+    const handleDeleteAll = async (id) => {
+
+        Swal.fire({
+            title: 'Tem certeza?',
+            text: 'Esta composição será removida permanentemente!',
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonText: 'Sim, remover',
+            cancelButtonText: 'Cancelar'
+        }).then(async (result) => {
+            if (result.isConfirmed) {
+                const response = await fetch(`${url}/composicao/${id}`, {
+                    method: 'DELETE',
+                    credentials: 'include',
+                    headers: {
+                        'Content-Type': 'application/json'
+                    }
+                })
+                if (response.ok) {
+                    onRemoveComposicao()
+                    Swal.fire('Removido!', 'Composicao familiar removido com sucesso', 'success')
+                } else {
+                    Swal.fire('Erro!', 'Não foi possível remover a composição familiar', 'error')
+                }
             }
+
         })
-        const data = await response.json()
-        console.log(data)
     }
 
+    const handleDelete = async (id) => {
+        Swal.fire({
+            title: 'Tem certeza?',
+            text: 'Este membro familiar será removido permanentemente!',
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonText: 'Sim, remover',
+            cancelButtonText: 'Cancelar'
+        }).then(async (result) => {
+            if (result.isConfirmed) {
+                const response = await fetch(`${url}/composicao_familiar/${id}`, {
+                    method: 'DELETE',
+                    credentials: 'include',
+                    headers: {
+                        'Content-Type': 'application/json'
+                    }
+                })
+                if (response.ok) {
+                    onRemoveMembro(id)
+                    Swal.fire('Removido!', 'Membro familiar removido com sucesso!', 'sucess')
+                } else {
+                    Swal.fire('Erro!', 'Não foi possível remover o membro familiar.', 'error')
+                }
+            }
+        })
+    }
 
     return (
         <div>
             <div className="d-flex justify-content-between">
                 <h3 className={`${styles['title-cadastro']}`}>Composição Familiar </h3>
-                <form>
-                    <button onClick={handleRemove} className="btn btn-danger btn-sm"> <BsTrash3Fill /> Apagar</button>
-                </form>
+                <button
+                    className="btn btn-sm btn-danger"
+                    onClick={() => { handleDeleteAll(composicaoFamiliar.id) }}>
+                    <BsTrashFill />
+                    Remover
+                </button>
             </div>
             {membro.map((el) => (
                 <div key={el.id}>
@@ -58,7 +105,20 @@ export const MembroId = ({ membro, composicaoFamiliar }) => {
                             <p className="col-12 col-md-3"><b>PCD: </b>{el.pcd}</p>
                         </div>
                         <div className="d-flex justify-content-between">
-                            <Link to={`/dashboard/update-dados-composicao-familiar/${el.id}`}><button className='btn btn-primary btn-sm'><HiOutlinePencilAlt />Editar</button></Link>
+                            <div>
+                                <Link to={`/dashboard/update-dados-composicao-familiar/${el.id}`}>
+                                    <button className='btn btn-primary btn-sm'><HiOutlinePencilAlt />
+                                        Editar
+                                    </button>
+                                </Link>
+
+                                <button
+                                    onClick={() => { handleDelete(el.id) }}
+                                    className="btn btn-sm btn-danger ms-2">
+                                    <BsTrashFill />
+                                    Apagar
+                                </button>
+                            </div>
                             <p className={`${styles['title-cadastro', 'data']} text-secondary`}>Última atualização: {formatarDataCriacaoAtualizacao(el.updatedAt)}</p>
                         </div>
                     </div>
